console - improve password recovery path (#4050)
Fixing behaviour, if the token is not found then the model is
incompletely prepared, and Java won't be able to provide the expected
form.

Preferring a redirect to passwordRecovery. This induces an extra 302
instead of getting the page directly, but sounds more logical, to get
back the whole form from the begining.

Also storing the "bad.token" error message in session to avoid having to
change the default configuration in the datadir (passwordRecovery does
not expect parameters), and keeping compatible with the current default
geOrchestra configuration.