console - setting the token as non-required parameter
This is not needed per se, but avoids to get a 500 error from spring,
which in turn triggers the "application unavailable" georchestra default
page, which is not relevant in this situation.

So allowing to have the token to be null, which will also redirect to
the password change page along with the same "bad token" error message.