fixing spring beans injection in rabbitMq specific xml defs
* refactor a bit RAbbitMqEventsListener to make use of lombok /
  autowired feature
* defining webmvc-config.xml as global configContextLocation, instead of
  using it only on the servlet.

It seems that the beans are "scoped" ; we have 2 different definitions
of contextConfigLocation, a first one which is "general" to the whole
webapp context, a second one specific to the spring servlet.

In the former version of the code, webmvc-config.xml was only loaded in
the scope of the servlet, and contains a bean (LogUtils) which is needed
in the first context, leading - in case of activating rabbitmq - to
unresolveable beans.

The modification here intends to have everything loaded at a webapp's
scope, to avoid dependency resolution hell.

Tests:
unfortunately, IT are useless here, as it is very servlet specific, so
tests have been mainly runtime, in a docker composition, considering
both servlet container in use in the community:

* in a jetty-based container: OK
* in a tomcat-based container: OK

In both cases, the flag `enableRabbitmqEvents` has been set to true &
false, which produced the following at runtime:

* `true`: `LogUtils` & `RabbitmqEventsListener` being instanciated only once
* `false`: Only `LogUtils` being instanciated.