console - gdpr export - manage classic ISO8601 date format
When the GDPR export feature has been implemented in 2019, it seems that
geonetwork was making use of a specific date format, which is not in use
anymore with newer versions, see here for the expected one:

https://github.com/georchestra/georchestra/blob/master/console/src/main/java/org/georchestra/console/ds/AccountGDPRDaoImpl.java#L57-L66

Checking in newer instances, it looks like we now have a classic
ISO_DATE_TIME format.

This commit suggests to first try to parse as ISO, then falls back onto
the former format. If still no success, a RuntimeException is raised.

Tests:
* UT added
* maven's `clean verify` OK