gwc - introduces an environment-aware xml configuration reader
this introduces a class which is basically a copy/paste from
the upstream XMLConfiguration class, which adds some converters
to the xstream configuration, so that `${...}` strings in the
XML are resolved against the environment variables.

Caveat: default implementation seems to also manage backup of the
configuration file. In this implementation, we forbid the possibility
to serialize back the configuration into the original XML file.

Caveat 2: This should be considered as a temporary solution, as
a PR on the XMLConfiguration class which would follow the GS / GWC
convention (e.g. honoring the `ALLOW_ENV_PARAMETRIZATION` java
property) should be issued upstream and hopefully merged, so that we
could stick to upstream / not custom-geOrchestra code.

Tests: 2 utests added, runtime tested by setting a description from an
env variable.