fix: making sure LDAP can listen on a UNIX socket
Basically creating the /var/run/slapd as it seems that the debian
package no longer provides it ? Or maybe slapd is unable to create it at
startup now but was able in the past.

Also taking the opportunity to allow slapd to listen on its UNIX socket
at runtime by modifying the CMD of the dockerfile. This will allow us to
access the LDAP configuration at runtime in the container.

Tests: runtime mainly. default image bootstrap is now ok, and I can
access the configuration via the UNIX socket.