fix: fixing the editOrgDetailsForm.jsp template in the console
Along with the following commit:
https://github.com/georchestra/georchestra/commit/a8d0733e8fab0aa56e8a2a437c4909cd094e1636#diff-b5a6c8a826b4bece338a05107c22595dc061f7475cb6b4a31924772fdfe276ecR92

A regression has been introduced to the console, making the servlet
container unable to compile the JSP template, because the attribute
name being used could not be resolved.

The fix is pretty trivial, but I took the opportunity to dig in how to
test modifications onto the JSP.

Tests:

Considering that the bug were introduced almost one year ago without
nobody ever noticing it since then (Thanks geo2france for the report), I
was looking for a way to avoid such issues in the future.

I first tried to use an integration test, but stumbled upon an issue
with testcontainers, and fixing things lead me to figure out that the
JSP templates were not even being processed anyway. So integration
tests were quite useless here, and probably depends on the servlet
container implementation you choose to deploy the webapp.

So I ended up introducing an maven plugin to precompile the JSP files,
which would fail in case of errors in the JSP files, which was my actual
goal. The suggested procedure to test the JSP is now documented in the
`README.md`.